used MoreExecutors.newDirectExecutorService() and refactored code
