+++ src/main/java/com/android/volley/toolbox/NetworkUtility.java
@@ -38,13 +38,7 @@ import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.SocketTimeoutException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.TreeSet;
 
 /**
  * Utility class for methods that are shared between {@link BasicNetwork} and {@link
@@ -52,51 +46,6 @@ import java.util.TreeSet;
  */
 public class NetworkUtility {
     private static final int SLOW_REQUEST_THRESHOLD_MS = 3000;
-
-    /**
-     * Combine cache headers with network response headers for an HTTP 304 response.
-     *
-     * <p>An HTTP 304 response does not have all header fields. We have to use the header fields
-     * from the cache entry plus the new ones from the response. See also:
-     * http://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.3.5
-     *
-     * @param responseHeaders Headers from the network response.
-     * @param entry The cached response.
-     * @return The combined list of headers.
-     */
-    static List<Header> combineHeaders(List<Header> responseHeaders, Cache.Entry entry) {
-        // First, create a case-insensitive set of header names from the network
-        // response.
-        Set<String> headerNamesFromNetworkResponse = new TreeSet<>(String.CASE_INSENSITIVE_ORDER);
-        if (!responseHeaders.isEmpty()) {
-            for (Header header : responseHeaders) {
-                headerNamesFromNetworkResponse.add(header.getName());
-            }
-        }
-
-        // Second, add headers from the cache entry to the network response as long as
-        // they didn't appear in the network response, which should take precedence.
-        List<Header> combinedHeaders = new ArrayList<>(responseHeaders);
-        if (entry.allResponseHeaders != null) {
-            if (!entry.allResponseHeaders.isEmpty()) {
-                for (Header header : entry.allResponseHeaders) {
-                    if (!headerNamesFromNetworkResponse.contains(header.getName())) {
-                        combinedHeaders.add(header);
-                    }
-                }
-            }
-        } else {
-            // Legacy caches only have entry.responseHeaders.
-            if (!entry.responseHeaders.isEmpty()) {
-                for (Map.Entry<String, String> header : entry.responseHeaders.entrySet()) {
-                    if (!headerNamesFromNetworkResponse.contains(header.getKey())) {
-                        combinedHeaders.add(new Header(header.getKey(), header.getValue()));
-                    }
-                }
-            }
-        }
-        return combinedHeaders;
-    }
 
     /** Logs requests that took over SLOW_REQUEST_THRESHOLD_MS to complete. */
     static void logSlowRequests(
@@ -113,26 +62,6 @@ public class NetworkUtility {
         }
     }
 
-    static Map<String, String> getCacheHeaders(Cache.Entry entry) {
-        // If there's no cache entry, we're done.
-        if (entry == null) {
-            return Collections.emptyMap();
-        }
-
-        Map<String, String> headers = new HashMap<>();
-
-        if (entry.etag != null) {
-            headers.put("If-None-Match", entry.etag);
-        }
-
-        if (entry.lastModified > 0) {
-            headers.put(
-                    "If-Modified-Since", HttpHeaderParser.formatEpochAsRfc1123(entry.lastModified));
-        }
-
-        return headers;
-    }
-
     static NetworkResponse getNetworkResponse(
             int statusCode, Request<?> request, long requestStart, List<Header> responseHeaders) {
         Cache.Entry entry = request.getCacheEntry();
@@ -145,7 +74,7 @@ public class NetworkUtility {
                     responseHeaders);
         }
         // Combine cached and response headers so the response will be complete.
-        List<Header> combinedHeaders = NetworkUtility.combineHeaders(responseHeaders, entry);
+        List<Header> combinedHeaders = HttpHeaderParser.combineHeaders(responseHeaders, entry);
         return new NetworkResponse(
                 HttpURLConnection.HTTP_NOT_MODIFIED,
                 entry.data,
